Refactor loadout service into a few ES modules
