+++ src/app/inventory/d2-stores.ts
@@ -17,8 +17,7 @@ import { getBuckets } from '../destiny2/d2-buckets';
 import { getDefinitions, D2ManifestDefinitions } from '../destiny2/d2-definitions';
 import { bungieNetPath } from '../dim-ui/BungieImage';
 import { reportException } from '../utils/exceptions';
-import { optimalLoadout } from '../loadout/loadout-utils';
-import { getLight } from '../loadout/loadout.service';
+import { optimalLoadout, getLight } from '../loadout/loadout-utils';
 import { resetIdTracker, processItems } from './store/d2-item-factory';
 import { makeVault, makeCharacter } from './store/d2-store-factory';
 import { NewItemsService } from './store/new-items';
