+++ src/app/loadout/LoadoutDrawerOptions.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { t } from 'app/i18next-t';
 import _ from 'lodash';
-import { Loadout } from './loadout.service';
+import { Loadout } from './loadout-types';
 import { router } from '../router';
 
 export default function LoadoutDrawerOptions(
