+++ src/app/storage/GDriveRevisions.tsx
@@ -8,7 +8,7 @@ import { SyncService } from './sync.service';
 import { UIViewInjectedProps } from '@uirouter/react';
 import { refreshIcon, AppIcon } from '../shell/icons';
 import { initSettings } from '../settings/settings';
-import { dimLoadoutService } from '../loadout/loadout.service';
+import { getLoadouts } from 'app/loadout/loadout-storage';
 
 interface State {
   revisions?: any;
@@ -164,7 +164,7 @@ class GDriveRevisionComponent extends React.Component<
         await SyncService.set(content, true);
         alert(t('Storage.ImportSuccess'));
         initSettings();
-        dimLoadoutService.getLoadouts(true);
+        getLoadouts(true);
         this.props.onRestoreSuccess();
       }
     }
